# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Consulta RNC - API Gratuita",
        "version": "1.2.5",
"description": "###  Herramienta de consulta de contribuyentes\n\nEste es un proyecto sencillo diseñado para facilitar el acceso a la información pública de contribuyentes en la República Dominicana.\n\n* **Desarrollado por:** RAXLOR SYSTEMS\n* **Límite de cortesía:** 2,500 peticiones cada 5 minutos\n* **Privacidad:** No guardamos registros de las consultas realizadas.",        "contact": {
            "name": "RAXLOR SYSTEMS",
            "url": "https://github.com/Raxlor"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor de Producción"
        }
    ],
    "tags": [
        {
            "name": "Consultas",
            "description": "Endpoints para la recuperación de datos de contribuyentes"
        }
    ],
    "paths": {
        "/api/v1/rnc/{rnc}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Consulta por identificador único",
                "description": "Obtiene la información detallada de un contribuyente utilizando su RNC o Cédula.",
                "parameters": [
                    {
                        "name": "rnc",
                        "in": "path",
                        "required": true,
                        "description": "RNC (9 dígitos) o Cédula (11 dígitos) sin guiones",
                        "schema": {
                            "type": "string",
                            "example": "101003723"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContribuyenteResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Identificador no registrado en la base de datos"
                    }
                }
            }
        },
        "/api/v1/buscar": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Búsqueda por razón social",
                "description": "Búsqueda de coincidencias basadas en el nombre comercial o razón social.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "description": "Término de búsqueda (mínimo 3 caracteres)",
                        "schema": {
                            "type": "string",
                            "example": "CERVECERIA NACIONAL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de la búsqueda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "exito": {
                                            "type": "boolean"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contribuyente"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contribuyente": {
                "type": "object",
                "properties": {
                    "rnc": {
                        "type": "string",
                        "example": "101003723"
                    },
                    "razon_social": {
                        "type": "string",
                        "example": "CERVECERIA NACIONAL DOMINICANA SA"
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "CERVECERIA NACIONAL DOMINICANA"
                    },
                    "actividad_economica": {
                        "type": "string",
                        "example": "ELABORACION DE CERVEZAS, BEBIDAS MALTEADAS Y MALTA"
                    },
                    "estado": {
                        "type": "string",
                        "example": "ACTIVO"
                    }
                }
            },
            "ContribuyenteResponse": {
                "type": "object",
                "properties": {
                    "exito": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/Contribuyente"
                    }
                }
            }
        }
    }
}